QML2 : Use real POIs icons
