+++ navit/gui/qt5_qml/CMakeLists.txt
@@ -5,4 +5,4 @@ set(CMAKE_AUTOMOC ON)
 
 qt5_add_resources(GUI_QT5_QML_QRC "gui_qt5_qml.qrc")
 
-module_add_library(gui_qt5_qml gui_qt5_qml.cpp ${GUI_QT5_QML_QRC} backend.cpp qml_map.cpp qml_poi.cpp)
+module_add_library(gui_qt5_qml gui_qt5_qml.cpp ${GUI_QT5_QML_QRC} backend.cpp qml_map.cpp qml_poi.cpp proxy.c)
