+++ navit/gui/qt5_qml/proxy.h
@@ -1,0 +1,1 @@
+char * get_icon(struct navit *nav, struct item *item);
